requirements.txt
streamlit
google-generativeai
speechrecognition
pillow
pymupdf
reportlab
google-api-core
